// src/pages/articles-page/model/types/articles-page-schema.ts
import { EntityState } from '@reduxjs/toolkit';
import { Article, ArticleView } from 'entities/article';

export interface ArticlesPageSchema extends EntityState<Article>{
  isLoading?: boolean
  error?: string
  view: ArticleView
  // pagination
  page: number
  limit?: number
  hasMore: boolean
}